Moved abstract classes to 'abstracts' directory
