+++ owlapy/abstracts/abstract_owl_reasoner.py
@@ -8,7 +8,7 @@ from owlapy.class_expression import OWLClassExpression
 from owlapy.class_expression import OWLClass
 from owlapy.owl_data_ranges import OWLDataRange
 from owlapy.owl_object import OWLEntity
-from owlapy.owl_ontology import OWLOntology
+from owlapy.abstracts.abstract_owl_ontology import OWLOntology
 from owlapy.owl_property import OWLObjectPropertyExpression, OWLDataProperty, OWLObjectProperty
 from owlapy.owl_individual import OWLNamedIndividual
 from owlapy.owl_literal import OWLLiteral
